requirements.txt
vk-api
spotipy
python-dotenv
colorama